Create README.md
In the gym measuring N×M meters built a modern attraction called "Left maze". To do this, on the floor of the gym at intervals of 1 meter drew lines parallel to the walls of the gym. Thus, the gym was divided into N×M cells. Then some of these cells were painted black. The attraction is that the participant is put in some cell of the gym and asked to run as fast as possible to some other cell. The following conditions apply:

The participant is forbidden to walk on the black cells.
Came in some kind of cage, the contestant can either go straight ahead or left or right (if in the appropriate direction the cell is not painted black): walk backwards and also diagonally is prohibited.
During the whole journey the participant is allowed to turn right (that is, to go from the current cell to the right relative to where he came to this cell) no more than K times.
In the initial cell, the participant can face in the direction in which he wants. Which side of the party resort in the destination cell is also not important.
It is known that the participant spends exactly 1 second to run from the cell to the next one. It is required to calculate the minimum time in which the participant will be able to reach the final cell.